Refactored Unit Tests
